improved validation when deleting comments from glossary SC#325

+++ mod/glossary/comment.php
@@ -80,7 +80,7 @@
         if (!$glossary->allowcomments && !has_capability('mod/glossary:managecomments', $context)) {
                 error('You can\'t delete comments in this glossary!');
             }
-        if ( $confirm ) {
+        if ( data_submitted() and $confirm ) {
             delete_records('glossary_comments','id', $cid);
 
             print_simple_box_start('center','40%', '#FFBBBB');
